Transformers_S13/requirements.txt
gradio
torch
transformers
pyyaml
sentencepiece
accelerate
